update mainPage components
