+++ src/pages/main/Activity.scss
