+++ src/pages/main/Board.scss
