+++ src/pages/main/ClubAwards.js
@@ -1,0 +1,9 @@
+function ClubAwards() {
+    return(
+        <div className="ClubAwards">
+            <h1>수상 실적</h1>
+        </div>
+    )
+}
+
+export default ClubAwards;
