+++ src/pages/main/ClubIntroduction.js
@@ -1,0 +1,38 @@
+import "./ClubIntroduction.scss";
+
+function clubIntroduction() {
+    return (
+      <div className="container">
+        <div className="Introduction">
+          <div className="title">동아리 소개</div>
+          <div className="clubIntros">
+            <div className="About">
+              <img src="free-animated-icon-checklist-6416398.gif" alt="아이콘1"></img>
+              <div className="label">About</div>
+              <div className="details">사법권은 법관으로 구성된 법원에 속한다. 행정각부의 설치·조직과 직무범위는 법률로 정한다. 국가는 전통문화의 계승·발전과 민족문화의 창달에 노력하여야 한다.</div>
+            </div>
+            <div className="Careers">
+              <img src="free-animated-icon-checklist-6416398.gif" alt="아이콘1"></img>
+              <div className="label">Careers</div>
+              <div className="details">사법권은 법관으로 구성된 법원에 속한다. 행정각부의 설치·조직과 직무범위는 법률로 정한다. 국가는 전통문화의 계승·발전과 민족문화의 창달에 노력하여야 한다.</div>
+  
+            </div>
+            <div className="Sammaru">
+              <img src="free-animated-icon-checklist-6416398.gif" alt="아이콘1"></img>
+              <div className="label">Sammaru</div>
+              <div className="details">사법권은 법관으로 구성된 법원에 속한다. 행정각부의 설치·조직과 직무범위는 법률로 정한다. 국가는 전통문화의 계승·발전과 민족문화의 창달에 노력하여야 한다.</div>
+  
+            </div>
+            <div className="Project">
+              <img src="free-animated-icon-checklist-6416398.gif" alt="아이콘1"></img>
+              <div className="label">Project</div>
+              <div className="details">사법권은 법관으로 구성된 법원에 속한다. 행정각부의 설치·조직과 직무범위는 법률로 정한다. 국가는 전통문화의 계승·발전과 민족문화의 창달에 노력하여야 한다.</div>
+  
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  }
+
+  export default clubIntroduction;
